write toml
